整理重複的程式碼
