+++ app/controllers/episodes_controller.rb
@@ -1,6 +1,6 @@
 class EpisodesController < ApplicationController
   before_action :find_episode, only: [:edit, :update, :show, :destroy]
-  before_action :podcaster_find_podcast, except: [:index]
+  before_action :find_podcast, except: [:index]
 
   def new
     @episode = Episode.new
@@ -43,5 +43,11 @@ class EpisodesController < ApplicationController
     @episode = Episode.find_by!(random_url: params[:id])
     rescue ActiveRecord::RecordNotFound
       redirect_to podcasts_path, notice: "找不到單集"
+  end
+
+  def find_podcast
+    @podcast = Podcast.find_by!(random_url: params[:podcast_id]) if @podcast.nil?
+    rescue ActiveRecord::RecordNotFound
+      redirect_to podcasts_path, notice: "找不到節目"
   end
 end
