+++ app/models/podcast.rb
@@ -34,16 +34,4 @@ class Podcast < ApplicationRecord
     Technology: "Technology"
   }
 
-  private
-  def generate_random_url
-    require 'securerandom'
-    new_random_url = SecureRandom.uuid
-    # 對Podcast做判斷式看看是否已經存在random_url
-    while Podcast.exists?(random_url: new_random_url)
-      new_random_url = SecureRandom.uuid
-    end
-
-    self.random_url = new_random_url
-  end
-
 end
