+++ app/views/episodes/new.html.erb
@@ -20,6 +20,5 @@
   <%= f.button :submit, "新增單集" %>
 <% end %>
 
-
 <br>
 <%= link_to "返回單集列表", podcast_path(@podcast.random_url) %>
