+++ app/views/pages/index.html.erb
@@ -5,6 +5,6 @@
 
   參考 <%= link_to "Soundon官網首頁", 'https://www.soundon.fm/', class: "btn bg-red-400 hover:bg-red-600", target: "_blank" %>
   <%= link_to "創作者", root_url(subdomain: "podcaster"), class: "btn btn-primary" %>
-  <%= link_to "收聽", browse_path(subdomain: "player"), class: "btn btn-primary" %>
+  <%= link_to "收聽", browse_url(subdomain: "player"), class: "btn btn-primary" %>
 
 </div>
