+++ app/views/podcasts/_form.html.erb
@@ -1,0 +1,23 @@
+<div class="create_table grid justify-center border-0 shadow rounded-r-none">
+  <%= simple_form_for podcast do |f| %>
+    <%= f.input :cover, as: :file, input_html:{class: "m-2 p-2" }, label: "上傳封面圖片" %>
+    <%= f.input :name, input_html:{class: "fill-form" }, label: "Podcast 節目名稱"%>
+    <%= f.input :artist, input_html:{class: "fill-form" }, label: "創作者名稱" %>
+    <%= f.input :email, input_html:{class: "fill-form" }, label: "Email" %>
+    <%= f.input :language, collection: ["繁體中文", "简体中文", "粵語", "English", "日本語", "한국어"], input_html:{class: "fill-form" }, label: "語言", placeholder: "Select a language" %>
+    <%= f.input :slug, input_html:{class: "fill-form" }, label: "好記短網址 https://soundbar.tk/" %>
+    <p class="text-purple-500 text-sm">必須為英文與數字，請勿使用特殊符號。</p>
+    <%= f.input :genres, collection: Podcast.genres.keys, input_html:{class: "fill-form" }, label: "類別", error: '請填寫類別' %>
+    <%= f.input :description, input_html:{class: "fill-form" }, label: "描述", placeholder: "描述這個 Podcast 所包含的內容。" %>
+    <%= f.input :subtitle, input_html:{class: "fill-form" }, label: "Subtitle" %>
+    <%= f.input :weblink, input_html:{class: "fill-form" }, label: "網站連結" %>
+    <%= f.input :copyright, input_html:{class: "fill-form" }, label: "版權宣告Copyright" %>
+    <%= f.input :explicit, input_html:{class: "m-2 p-2" }, label: "兒童不宜" %>
+    <%= f.input :status, collection: ["草稿", "發布"], input_html:{class: "fill-form-status" }, label: "狀態" %>
+
+    <div class='flex items-center justify-center  md:gap-8 gap-4 pt-5 pb-5'>
+      <button class='btn bg-gray-500 hover:bg-gray-700'><%= link_to "返回節目列表", podcasts_path %></button>
+      <%= f.button :submit, "儲存", class: "btn-primary btn" %>
+    </div>
+  <% end %>
+</div>
